+++ src/main/scala/xerial/sbt/Pack.scala
@@ -34,6 +34,10 @@ object Pack extends sbt.Plugin {
     override def toString = "%s:%s:%s%s".format(org, name, revision, classifierSuffix)
 
     def jarName = "%s-%s%s.jar".format(name, revision, classifierSuffix)
+
+    def fullJarName = "%s.%s-%s%s.jar".format(org, name, revision, classifierSuffix)
+
+    def noVersionJarName = "%s.%s%s.jar".format(org, name, classifierSuffix)
   }
 
   private implicit def moduleEntryOrdering = Ordering.by[ModuleEntry, (String, String, String, Option[String])](m => (m.org, m.name, m.revision, m.classifier))
@@ -42,7 +46,7 @@ object Pack extends sbt.Plugin {
 
   val pack = taskKey[File]("create a distributable package of the project")
   val packDir = settingKey[String]("pack-dir")
-  
+
   val packBashTemplate = settingKey[String]("template file for bash scripts - defaults to pack's out-of-the-box template for bash")
   val packBatTemplate = settingKey[String]("template file for bash scripts - defaults to pack's out-of-the-box template for bat")
   val packMakeTemplate = settingKey[String]("template file for bash scripts - defaults to pack's out-of-the-box template for make")
@@ -62,10 +66,10 @@ object Pack extends sbt.Plugin {
   val packAllUnmanagedJars = taskKey[Seq[Classpath]]("all unmanaged jar files")
   val packJvmOpts = SettingKey[Map[String, Seq[String]]]("pack-jvm-opts")
   val packExtraClasspath = SettingKey[Map[String, Seq[String]]]("pack-extra-classpath")
-  val packPreserveOriginalJarName = SettingKey[Boolean]("pack-preserve-jarname", "preserve the original jar file names. default = false")
-  
+  val packJarNameConvention = SettingKey[String]("pack-jarname-convention", "default: (artifact name)-(version).jar; original: original JAR name; full: (organization).(artifact name)-(version).jar; no-version: (organization).(artifact name).jar")
+
   val DEFAULT_RESOURCE_DIRECTORY = "src/pack"
-  
+
   lazy val packSettings = Seq[Def.Setting[_]](
     packDir := "pack",
     packBashTemplate := "/xerial/sbt/template/launch.mustache",
@@ -74,14 +78,14 @@ object Pack extends sbt.Plugin {
     packMain := Map.empty,
     packExclude := Seq.empty,
     packMacIconFile := "icon-mac.png",
-    packResourceDir := Seq.empty,
+    packResourceDir := Seq(DEFAULT_RESOURCE_DIRECTORY),
     packJvmOpts := Map.empty,
     packExtraClasspath := Map.empty,
     packAllClasspaths <<= (thisProjectRef, buildStructure) flatMap getFromAllProjects(dependencyClasspath.task in Runtime),
     packAllUnmanagedJars <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(unmanagedJars.task in Compile),
     packLibJars <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(packageBin.task in Runtime),
     packUpdateReports <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(update.task),
-    packPreserveOriginalJarName := false,
+    packJarNameConvention := "default",
     packGenerateWindowsBatFile := true,
     (mappings in pack) := Seq.empty,
     pack := {
@@ -117,7 +121,12 @@ object Pack extends sbt.Plugin {
       // Copy dependent jars
       out.log.info("project dependencies:\n" + dependentJars.keys.mkString("\n"))
       for ((m, f) <- dependentJars) {
-        val targetFileName = if (packPreserveOriginalJarName.value) m.originalFileName else m.jarName
+        val targetFileName = packJarNameConvention.value match {
+          case "original" => m.originalFileName
+          case "full" => m.fullJarName
+          case "no-version" => m.noVersionJarName
+          case _ => m.jarName
+        }
         IO.copyFile(f, libDir / targetFileName, true)
       }
 
@@ -184,8 +193,8 @@ object Pack extends sbt.Plugin {
       }
 
       // Copy resources in src/pack folder
-      val binScriptsDir = base / DEFAULT_RESOURCE_DIRECTORY / "bin"
-      val otherResourceDirs = Seq(binScriptsDir) ++ packResourceDir.value.map( dir => base / dir )
+      val otherResourceDirs = packResourceDir.value.map( dir => base / dir )
+      val binScriptsDir = otherResourceDirs.map(_ / "bin").filter(_.exists)
       out.log.info(s"packed resource directories = ${otherResourceDirs.mkString(",")}")
 
       def linkToScript(name: String) =
@@ -193,7 +202,7 @@ object Pack extends sbt.Plugin {
 
       // Create Makefile
       val makefile = {
-        val additionalScripts = (Option(binScriptsDir.listFiles) getOrElse Array.empty).map(_.getName)
+        val additionalScripts = binScriptsDir.flatMap(_.listFiles).map(_.getName)
         val symlink = (mainTable.keys ++ additionalScripts).map(linkToScript).mkString("\n")
         val globalVar = Map("PROG_NAME" -> name.value, "PROG_SYMLINK" -> symlink)
         engine.layout(packMakeTemplate.value, globalVar)